UserActionLog

+++ Schemas/CreateAccountModel.cs
@@ -2,5 +2,5 @@
 
 public class CreateAccountModel
 {
-    public int Balance { get; set; }
+    public decimal Balance { get; set; }
 }
